# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RuiQi 应用防火墙管理系统 API",
        "title": "RuiQi-WAF API",
        "termsOfService": "https://github.com/HUAHUAI23/RuiQi",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/HUAHUAI23/RuiQi",
            "email": "huahua1319873800@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:2333",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/site": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有站点配置列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点管理"
                ],
                "summary": "获取站点列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取站点列表成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SiteListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的站点配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点管理"
                ],
                "summary": "创建新站点",
                "parameters": [
                    {
                        "description": "站点信息",
                        "name": "site",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSiteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "站点创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SiteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "409": {
                        "description": "域名和端口组合已存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
        "/api/v1/site/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取站点详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点管理"
                ],
                "summary": "获取单个站点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "站点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取站点详情成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SiteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "404": {
                        "description": "站点不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新指定站点的配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点管理"
                ],
                "summary": "更新站点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "站点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "站点更新信息",
                        "name": "site",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSiteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "站点更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SiteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "404": {
                        "description": "站点不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "409": {
                        "description": "域名和端口组合已被其他站点使用",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定的站点配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点管理"
                ],
                "summary": "删除站点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "站点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "站点删除成功",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessResponseNoData"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "404": {
                        "description": "站点不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BackendDTO": {
            "type": "object",
            "required": [
                "servers"
            ],
            "properties": {
                "servers": {
                    "description": "服务器列表，至少需要一个服务器",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.ServerDTO"
                    }
                }
            }
        },
...
        "dto.CertificateDTO": {
            "type": "object",
            "required": [
                "certName",
                "expireDate",
                "fingerPrint",
                "issuerName",
                "privateKey",
                "publicKey"
            ],
            "properties": {
                "certName": {
                    "description": "证书名称",
                    "type": "string",
                    "example": "my-cert"
                },
                "expireDate": {
                    "description": "过期时间",
                    "type": "string"
                },
                "fingerPrint": {
                    "description": "证书指纹",
                    "type": "string"
                },
                "issuerName": {
                    "description": "颁发机构",
                    "type": "string",
                    "example": "Let's Encrypt"
                },
                "privateKey": {
                    "description": "私钥内容",
                    "type": "string"
                },
                "publicKey": {
                    "description": "公钥内容",
                    "type": "string"
                }
            }
        },
...
        "dto.CreateSiteRequest": {
            "description": "创建站点的请求参数",
            "type": "object",
            "required": [
                "backend",
                "domain",
                "listenPort",
                "name"
            ],
            "properties": {
                "activeStatus": {
                    "description": "站点状态",
                    "type": "boolean",
                    "example": true
                },
                "backend": {
                    "description": "后端服务器配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.BackendDTO"
                        }
                    ]
                },
                "certificate": {
                    "description": "证书信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.CertificateDTO"
                        }
                    ]
                },
                "domain": {
                    "description": "域名",
                    "type": "string",
                    "example": "example.com"
                },
                "enableHTTPS": {
                    "description": "是否启用HTTPS",
                    "type": "boolean",
                    "example": false
                },
                "listenPort": {
                    "description": "监听端口",
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 1,
                    "example": 8080
                },
                "name": {
                    "description": "站点名称",
                    "type": "string",
                    "example": "my-site"
                },
                "wafEnabled": {
                    "description": "是否启用WAF",
                    "type": "boolean",
                    "example": false
                },
                "wafMode": {
                    "description": "WAF模式",
                    "type": "string",
                    "enum": [
                        "protection",
                        "observation"
                    ],
                    "example": "observation"
                }
            }
        },
...
        "dto.SiteListResponse": {
            "description": "站点列表响应",
            "type": "object",
            "properties": {
                "items": {
                    "description": "站点列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Site"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "dto.SiteResponse": {
            "description": "站点信息响应",
            "type": "object",
            "properties": {
                "activeStatus": {
                    "description": "站点是否激活",
                    "type": "boolean"
                },
                "backend": {
                    "description": "后端服务器配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Backend"
                        }
                    ]
                },
                "certificate": {
                    "description": "证书信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Certificate"
                        }
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "domain": {
                    "description": "域名，如 a.com",
                    "type": "string"
                },
                "enableHTTPS": {
                    "description": "是否启用HTTPS",
                    "type": "boolean"
                },
                "id": {
                    "description": "站点ID",
                    "type": "string"
                },
                "listenPort": {
                    "description": "监听端口，如 9000",
                    "type": "integer"
                },
                "name": {
                    "description": "站点名称",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "wafEnabled": {
                    "description": "是否启用WAF",
                    "type": "boolean"
                },
                "wafMode": {
                    "description": "WAF防护模式",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.WAFMode"
                        }
                    ]
                }
            }
        },
...
        "dto.UpdateSiteRequest": {
            "description": "更新站点的请求参数",
            "type": "object",
            "properties": {
                "activeStatus": {
                    "description": "站点状态",
                    "type": "boolean",
                    "example": true
                },
                "backend": {
                    "description": "后端服务器配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.BackendDTO"
                        }
                    ]
                },
                "certificate": {
                    "description": "证书信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.CertificateDTO"
                        }
                    ]
                },
                "domain": {
                    "description": "域名",
                    "type": "string",
                    "example": "example.com"
                },
                "enableHTTPS": {
                    "description": "是否启用HTTPS",
                    "type": "boolean",
                    "example": false
                },
                "listenPort": {
                    "description": "监听端口",
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 1,
                    "example": 8080
                },
                "name": {
                    "description": "站点名称",
                    "type": "string",
                    "example": "my-site"
                },
                "wafEnabled": {
                    "description": "是否启用WAF",
                    "type": "boolean",
                    "example": false
                },
                "wafMode": {
                    "description": "WAF模式",
                    "type": "string",
                    "enum": [
                        "protection",
                        "observation"
                    ],
                    "example": "observation"
                }
            }
        },
...
        "model.ErrResponse": {
            "description": "错误的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "参数错误"
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "model.ErrResponseDontShowError": {
            "description": "错误的API响应标准格式,不展示 error",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
        "model.Site": {
            "type": "object",
            "properties": {
                "activeStatus": {
                    "description": "站点是否激活",
                    "type": "boolean"
                },
                "backend": {
                    "description": "后端服务器配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Backend"
                        }
                    ]
                },
                "certificate": {
                    "description": "证书信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Certificate"
                        }
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "domain": {
                    "description": "域名，如 a.com",
                    "type": "string"
                },
                "enableHTTPS": {
                    "description": "是否启用HTTPS",
                    "type": "boolean"
                },
                "id": {
                    "description": "站点ID",
                    "type": "string"
                },
                "listenPort": {
                    "description": "监听端口，如 9000",
                    "type": "integer"
                },
                "name": {
                    "description": "站点名称",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "wafEnabled": {
                    "description": "是否启用WAF",
                    "type": "boolean"
                },
                "wafMode": {
                    "description": "WAF防护模式",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.WAFMode"
                        }
                    ]
                }
            }
        },
        "model.SuccessResponse": {
            "description": "成功的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "model.SuccessResponseNoData": {
            "description": "成功的API响应标准格式,没有 data",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 Bearer {token} 格式进行身份验证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}